Extract Response class
