+++ build.gradle
@@ -22,7 +22,8 @@ dependencies {
     api libs.org.springframework.boot.spring.boot.starter.data.jdbc
     api libs.org.springframework.boot.spring.boot.starter.jdbc
     api libs.org.springframework.boot.spring.boot.starter.web
-    api libs.org.projectlombok.lombok
+    compileOnly 'org.projectlombok:lombok:1.18.30'
+    annotationProcessor 'org.projectlombok:lombok:1.18.30'
     testImplementation libs.org.springframework.boot.spring.boot.starter.test
 }
