+++ src/main/java/com/skypro/StarBank/controller/RecommendationController.java
@@ -1,7 +1,8 @@
 package com.skypro.StarBank.controller;
 
-import com.skypro.StarBank.model.RecommendationDTO;
+import com.skypro.StarBank.dto.response.RecommendationDTO;
 import com.skypro.StarBank.service.RecommendationService;
+import com.skypro.StarBank.dto.response.Response;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -11,6 +12,7 @@ import java.util.List;
 public class RecommendationController {
 
     private final RecommendationService recommendationService;
+    
 
     public RecommendationController(RecommendationService recommendationService) {
         this.recommendationService = recommendationService;
@@ -20,18 +22,5 @@ public class RecommendationController {
     public Response getRecommendations(@PathVariable("user_id") String userId) {
         List<RecommendationDTO> recommendations = recommendationService.getRecommendations(userId);
         return new Response(userId, recommendations);
-    }
-
-    private static class Response {
-        private String user_id;
-        private List<RecommendationDTO> recommendations;
-
-        public Response(String user_id, List<RecommendationDTO> recommendations) {
-            this.user_id = user_id;
-            this.recommendations = recommendations;
-        }
-
-        public String getUser_id() { return user_id; }
-        public List<RecommendationDTO> getRecommendations() { return recommendations; }
     }
 }
