+++ src/main/java/com/skypro/StarBank/model/ProductCatalog.java
@@ -2,6 +2,8 @@ package com.skypro.StarBank.model;
 
 import org.springframework.stereotype.Component;
 
+import com.skypro.StarBank.dto.response.RecommendationDTO;
+
 import java.util.List;
