+++ src/main/java/com/skypro/StarBank/rules/RecommendationRuleSet.java
@@ -1,8 +1,8 @@
 package com.skypro.StarBank.rules;
 
-import com.skypro.StarBank.model.RecommendationDTO;
-
 import java.util.Optional;
+
+import com.skypro.StarBank.dto.response.RecommendationDTO;
 
 public interface RecommendationRuleSet {
     Optional<RecommendationDTO> check(String userId);
